Update git_basics.md
Hello,

I found an error, if I directly type 'git status' to the terminal in VS Code it's in the 'home' drive so therefore fails. This change, corrects the directory and then 'git status' works without error.

Hopefully this is approved, it'll be my first proper 'commit' and it's a small contribution to this amazing content I'm working through. Thank you.